+++ test/cli-validator/tests/circular-refs-validator.test.js
@@ -1,6 +1,9 @@
 const commandLineValidator = require('../../../src/cli-validator/runValidator');
 const circularRefsValidator = require('../../../src/cli-validator/utils/circular-references-ibm');
-const { getCapturedText } = require('../../test-utils');
+const {
+  getCapturedText,
+  getMessageAndPathFromCapturedText
+} = require('../../test-utils');
 
 describe('cli tool - test circular reference module', function() {
   it('should correctly validate a file with circular references', async function() {
@@ -86,5 +89,36 @@ describe('cli tool - test circular reference module', function() {
         'Swagger object should not contain circular references.'
       )
     ).toEqual(true);
+  });
+
+  it('should not fail when a multi-file spec contains circular references across files', async function() {
+    const consoleSpy = jest.spyOn(console, 'log').mockImplementation(() => {});
+
+    const program = {};
+    program.args = ['./test/cli-validator/mockFiles/multi-file-spec/main.yaml'];
+    program.default_mode = true;
+    program.ruleset = 'test/spectral/mockFiles/mockConfig/extends-default.yaml';
+
+    const exitCode = await commandLineValidator(program);
+
+    const capturedText = getCapturedText(consoleSpy.mock.calls);
+    consoleSpy.mockRestore();
+
+    expect(exitCode).toEqual(0);
+
+    const messages = getMessageAndPathFromCapturedText(
+      'Swagger object should not contain circular references.',
+      capturedText
+    );
+
+    expect(messages.length).toBe(1);
+
+    expect(messages[0][0].get('Message')).toEqual(
+      'Swagger object should not contain circular references.'
+    );
+
+    expect(messages[0][1].get('Path')).toEqual(
+      'paths./circular_example.get.responses.200.content.application/json.schema.properties.city.properties.primary_team.properties.location.properties'
+    );
   });
 });